feat: Added support for A records/IPv4
This commit adds support for A records and IPv4 addresses.

It also improves error messages, both for the internal log as well as
the HTTP response. This was especially needed as there are now a number
of different error reasons.

Some query parameters are not optional, so now it is not necessary
anymore to provide them even though they are empty.