feat: Added tests for server behaviour
Tests for server behaviour were added. This required some extensive
rewrites, so new dependencies were needed and some function signatures
changed. As there are no stability guarantees right now, this is not
technically a breaking change, but to signal this change nontheless the
minor version number is increased.

There is a new MockProvider that mimicks the behaviour of a Provider
implementation. To facilitate this, the configured providers now do not
save the UserConfig or anything like API tokens; instead, those
parameters have to be passed on each call, allowing Provider instances
to be reused. The configured Providers are managed via Rocket state.

Additionally, the default log level is set to Warn instead of Debug,
thus removing many of the very verbose log messages from Rocket.